+++ tsmap.go
@@ -4,31 +4,31 @@ package tsmap
 
 import "sync"
 
-type TsMap[K comparable, V any] struct {
+type ThreadSafeMap[K comparable, V any] struct {
 	coreMap map[K]V
-	mtx     sync.RWMutex
+	mu      sync.RWMutex
 }
 
-func NewTsMap[K comparable, V any]() TsMap[K, V] {
-	return TsMap[K, V]{coreMap: map[K]V{}}
+func NewThreadSafeMap[K comparable, V any]() ThreadSafeMap[K, V] {
+	return ThreadSafeMap[K, V]{coreMap: map[K]V{}}
 }
 
-func (m *TsMap[K, V]) Get(key K) (value V, exist bool) {
-	m.mtx.RLock()
-	defer m.mtx.RUnlock()
+func (m *ThreadSafeMap[K, V]) Get(key K) (value V, exist bool) {
+	m.mu.RLock()
+	defer m.mu.RUnlock()
 	value, exist = m.coreMap[key]
 	return value, exist
 }
 
-func (m *TsMap[K, V]) Set(key K, value V) {
-	m.mtx.Lock()
-	defer m.mtx.Unlock()
+func (m *ThreadSafeMap[K, V]) Set(key K, value V) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
 	m.coreMap[key] = value
 }
 
-func (m *TsMap[K, V]) Pop(key K) (value V, exist bool) {
-	m.mtx.Lock()
-	defer m.mtx.Unlock()
+func (m *ThreadSafeMap[K, V]) Pop(key K) (value V, exist bool) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
 	value, exist = m.coreMap[key]
 	if exist {
 		delete(m.coreMap, key)
@@ -36,9 +36,9 @@ func (m *TsMap[K, V]) Pop(key K) (value V, exist bool) {
 	return value, exist
 }
 
-func (m *TsMap[K, V]) Delete(key K) (deleted bool) {
-	m.mtx.Lock()
-	defer m.mtx.Unlock()
+func (m *ThreadSafeMap[K, V]) Delete(key K) (deleted bool) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
 	if _, exists := m.coreMap[key]; exists {
 		delete(m.coreMap, key)
 		return true
@@ -47,22 +47,30 @@ func (m *TsMap[K, V]) Delete(key K) (deleted bool) {
 	}
 }
 
-func (m *TsMap[K, V]) ForEach(f func(value V)) {
-	m.mtx.Lock()
-	defer m.mtx.Unlock()
+func (m *ThreadSafeMap[K, V]) DeleteMultiple(keys ...K) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	for _, key := range keys {
+		delete(m.coreMap, key)
+	}
+}
+
+func (m *ThreadSafeMap[K, V]) ForEach(f func(value V)) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
 	for _, value := range m.coreMap {
 		f(value)
 	}
 }
 
-func (m *TsMap[K, V]) Size() int {
-	m.mtx.RLock()
-	defer m.mtx.RUnlock()
+func (m *ThreadSafeMap[K, V]) Size() int {
+	m.mu.RLock()
+	defer m.mu.RUnlock()
 	return len(m.coreMap)
 }
 
-func (m *TsMap[K, V]) Flush() {
-	m.mtx.Lock()
-	defer m.mtx.Unlock()
+func (m *ThreadSafeMap[K, V]) Flush() {
+	m.mu.Lock()
+	defer m.mu.Unlock()
 	m.coreMap = make(map[K]V)
 }