+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # tsmap
 
+## [v0.2.0]
+
+- renameed main struct
+- added method `DeleteMultiple`
+
 ## [v0.1.0]
 
 - init